Add very basic version of first-class references
Currently, only 1 reference of a single type may be in scope. Multiple
references of the same type will clobber eachother on writes, and
references with different types will cause the function containing them
not to typecheck.

Global references are unaffected.

The 'ref' type is also not yet available for function arguments, so
references are currently restricted to the context of the function that
creates them.