Make module names optional, leave modules fold through unnamed modules
This let us use unnamed modules to make up an actual module out of
chunks. In particular, this makes it much easier to give the right
results for e.g.

module M = struct
  let a = 15
end

module A = struct
  let a = 20
  let x = a
  open M
  let y = a
end

module B = struct
  open M
  let x = a
  let a = 20
  let y = a
end

Currently, module B is interpreted incorrectly: the value a is resolved
to M.a in B.y when it should have been shadowed by the definition of
B.a.