Add an initial version of a union effect
This creates a union Effect.t, which carries multiple other effects. By
design, it allows multiple effects in its output list to resolve to a
single effect in its input list, so that type resolution doesn't break
state (e.g. union [state Z] [state A; state B] when A = B = nat).

Currently, the utility functions require that the union effect is first
in the effects list, for simplicity.