Infer the type of a matched expression from the case expressions
Coq needs to be told or infer what the type variables are when they
occur, so we need to know the actual type of a typed expression, and
then we can coerce those unknown type variables to unit.

We can't get this information from the OCaml compiler, since its  type
unification system doesn't unify type variables in matches, so we have
to do this work independently.